refactor: 分离Controller生成器逻辑
